Create invitee controller
