+++ src/index.ts
@@ -1,7 +1,7 @@
 import express from "express";
 import dotenv from "dotenv";
 
-import inviteeRouter from "./routes/invitee";
+import inviteeRouter from "./routes/inviteeRouter";
 
 dotenv.config({ path: "./src/config/config.env" });
 const app = express();
